Code audit wip
