+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/languages/ImportLanguages.kt
@@ -1,11 +1,10 @@
-package org.wycliffeassociates.otter.common.domain
+package org.wycliffeassociates.otter.common.domain.languages
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.core.JsonFactory
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import io.reactivex.Completable
-import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import org.wycliffeassociates.otter.common.data.model.Language
 import org.wycliffeassociates.otter.common.persistence.repositories.ILanguageRepository
@@ -18,7 +17,7 @@ class ImportLanguages(val file: File, val languageRepo: ILanguageRepository) {
             val mapper = ObjectMapper(JsonFactory())
             mapper.registerModule(KotlinModule())
             val languages = file.bufferedReader().use {
-                mapper.readValue(it, Array<Door43Lanuage>::class.java)
+                mapper.readValue(it, Array<Door43Language>::class.java)
             }
             languageRepo.insertAll(languages.toList().map { it.toLanguage() }).blockingGet()
         }.subscribeOn(Schedulers.io())
@@ -26,7 +25,7 @@ class ImportLanguages(val file: File, val languageRepo: ILanguageRepository) {
 }
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-private data class Door43Lanuage(
+private data class Door43Language(
     val pk: Int, // id
     val lc: String, // slug
     val ln: String, // name
