+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/ImportResourceContainer.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.domain
+package org.wycliffeassociates.otter.common.domain.resourcecontainer
 
 import io.reactivex.Completable
 import io.reactivex.Single
