+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/OtterResourceContainerConfig.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.domain
+package org.wycliffeassociates.otter.common.domain.resourcecontainer
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
