+++ src/main/kotlin/org/wycliffeassociates/otter/common/persistence/repositories/ICollectionRepository.kt
@@ -11,6 +11,9 @@ import org.wycliffeassociates.resourcecontainer.ResourceContainer
 
 interface ICollectionRepository : IRepository<Collection> {
     fun insert(collection: Collection): Single<Int>
+    fun getRootProjects(): Single<List<Collection>>
+    fun getRootSources(): Single<List<Collection>>
+    fun getSource(project: Collection): Maybe<Collection>
     fun getBySlugAndContainer(slug: String, container: ResourceMetadata): Maybe<Collection>
     fun getChildren(collection: Collection): Single<List<Collection>>
     fun updateSource(collection: Collection, newSource: Collection): Completable
